made it more organized for adding the json from firebase
